requirements.txt
flask
yahoo-finance
flask-sqlalchemy
sqlalchemy-utils
requests
bs4
python-dateutil
scipy
numpy
